Working on error handling involving Youtube ids.
Not perfect, but it will work for now. The original play was to remove
all ids that the player errored on, but this proved problematic,  as
occasionally, it would double error, removing the video with the error,
and the next one. This is still flawed, as not all video ids are valid
yet, but, again, this will work for the time being.