Change from Sortable() to McPants/jquery.shapeshift.
The change was pretty necessary, as some major issues exist when
attempting to interface between Bootstrap Grid and jQuery sortable. That
said, trying to reinvent the wheel for that element was not helping in
this situation. There are still kinks to work out, but this is a far
better state than that which we were in. That said, this broke a rule,
to a degree, I set for myself in not having to have any modules
interface directly with the base page, which is now necessary to reset
shapeshift() on an expansion.

-Also, resizing of pretty much everything, mainly to support consistent
height between expanded and inactive modules.